+++ src/Traits/AsSelfAwareableEnumTrait.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Phil\Enums\Traits;
+
+trait AsSelfAwareableEnumTrait
+{
+    /**
+     * Determine whether the enum is pure.
+     */
+    public static function isPure(): bool
+    {
+        return !self::isBacked();
+    }
+
+    /**
+     * Determine whether the enum is backed.
+     */
+    public static function isBacked(): bool
+    {
+        /** @phpstan-ignore function.impossibleType */
+        return is_subclass_of(self::class, BackedEnum::class);
+    }
+
+    /**
+     * Determine whether the enum is backed by integer.
+     */
+    public static function isBackedByInteger(): bool
+    {
+        return (string) (new ReflectionEnum(self::class))->getBackingType() === 'int';
+    }
+
+    /**
+     * Determine whether the enum is backed by string.
+     */
+    public static function isBackedByString(): bool
+    {
+        return (string) (new ReflectionEnum(self::class))->getBackingType() === 'string';
+    }
+
+}