+++ CHANGELOG.md
@@ -29,6 +29,11 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 ### Added
 - Label Meta Property
 - Add equalsOneOf and notEqualsOneOf in AsComparableEnumTrait
+- Add self-awareness
+  * isPure
+  * isBacked
+  * isBackedByInteger
+  * isBackedByString
 
 ### Changed
 - Nothing